Move mdx configuration to its own file
