+++ src/posts/test.11tydata.js
@@ -1,0 +1,4 @@
+export default {
+  title: "Test",
+  tags: ["post"],
+};
